requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
# For Groq
groq
# For Google Gemini
google-generativeai